test new process of uploading files (large and small) and creating new versions
